requirements.txt
inference-gpu[yolo-world]==0.9.12
supervision==0.19.0rc3
gradio==4.19.0